[FIX] PWA 설치 상태 감지 및 설치 프롬프트 전역 관리 (#1661)
* fix: PWA 설치 상태를 전역으로 관리 #1660

* fix: 앱 최상단에서 PWA 설치 상태 감지 #1660

* fix: PWA 재설치 가능 시 설치 이력 초기화 #1660
