+++ frontend/src/common/hooks/usePWAInstall.ts
@@ -1,6 +1,4 @@
-import { useCallback, useEffect, useState } from 'react';
-
-import { captureSentryError } from '../utils/captureSentryError';
+import { createContext, useContext } from 'react';
 
 export interface BeforeInstallPromptEvent extends Event {
   prompt: () => Promise<void>;
@@ -10,70 +8,22 @@ export interface BeforeInstallPromptEvent extends Event {
   }>;
 }
 
-interface UsePWAInstallResult {
+export interface UsePWAInstallResult {
   canInstall: boolean;
+  hasInstalledBefore: boolean;
   promptInstall: () => Promise<void>;
   resetInstallPrompt: () => void;
 }
 
+export const PWAInstallContext = createContext<UsePWAInstallResult>({
+  canInstall: false,
+  hasInstalledBefore: false,
+  promptInstall: async () => {},
+  resetInstallPrompt: () => {},
+});
+
 const usePWAInstall = (): UsePWAInstallResult => {
-  const [installPromptEvent, setInstallPromptEvent] =
-    useState<BeforeInstallPromptEvent | null>(null);
-
-  const resetInstallPrompt = useCallback(() => {
-    setInstallPromptEvent(null);
-  }, []);
-
-  const promptInstall = useCallback(async (): Promise<void> => {
-    if (!installPromptEvent) {
-      return;
-    }
-
-    try {
-      await installPromptEvent.prompt();
-      await installPromptEvent.userChoice;
-      resetInstallPrompt();
-    } catch (error) {
-      console.error('[PWA] 설치 프롬프트 실행 실패:', error);
-      captureSentryError({
-        error,
-        level: 'warning',
-        feature: 'pwa',
-        step: 'install-prompt-execute',
-      });
-      resetInstallPrompt();
-    }
-  }, [installPromptEvent, resetInstallPrompt]);
-
-  useEffect(() => {
-    const handleBeforeInstallPrompt = (event: Event) => {
-      const beforeInstallPromptEvent = event as BeforeInstallPromptEvent;
-
-      event.preventDefault();
-      setInstallPromptEvent(beforeInstallPromptEvent);
-    };
-
-    const handleAppInstalled = () => {
-      resetInstallPrompt();
-    };
-
-    window.addEventListener('beforeinstallprompt', handleBeforeInstallPrompt);
-    window.addEventListener('appinstalled', handleAppInstalled);
-
-    return () => {
-      window.removeEventListener(
-        'beforeinstallprompt',
-        handleBeforeInstallPrompt,
-      );
-      window.removeEventListener('appinstalled', handleAppInstalled);
-    };
-  }, [resetInstallPrompt]);
-
-  return {
-    canInstall: installPromptEvent !== null,
-    promptInstall,
-    resetInstallPrompt,
-  };
+  return useContext(PWAInstallContext);
 };
 
 export default usePWAInstall;
